backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-socketio==5.11.0
python-engineio==4.9.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-dotenv
pydantic
pydantic-settings
asyncpg

# LangChain ecosystem
langchain
langchain-google-genai
langchain-ollama
langchain-community

# Vector stores and embeddings
faiss-cpu


# Document processing
pypdf2
pandas
python-docx

# Database and ORM
sqlalchemy
alembic
psycopg2-binary

# Utilities
httpx
aiofiles
Pillow
numpy
scikit-learn
cloudinary
